guix: packages: Extend bag-build to support gexp.
Rationale: The goal of this proposition was to enable guix users to
edit build-system lower builders on-the-fly, which is currently not
possible.  This is done by keeping a higher-level representation of
the builder - a gexp rather than a derivation - as the bag-build
field.

* guix/build-system.scm (<bag>): Update comment.
* guix/packages.scm
(bag->derivation): Handle gexp bag-build. Rename current procedure to...
(bag-builder):... this procedure.
(bag->cross-derivation): Rename procedure to...
(bag-cross-builder):... this procedure.

Change-Id: I56c5a9dab9954307f95b29eab5e02ee058271684