gnu: go-github-com-hodgesds-perf-utils: Skip more tests on arm.
I see these failing when built on Honeycomb build machines:

  === RUN   TestDataTLB
      utils_test.go:234: no such file or directory
  --- FAIL: TestDataTLB (0.00s)
  === RUN   TestInstructionTLB
      utils_test.go:246: no such file or directory
  --- FAIL: TestInstructionTLB (0.00s)

It could be something with the system configuration but just disable the tests
for now. This is working towards getting the prometheus-node-exporter package
building again on aarch64-linux.

Note that there's even more test failures when trying to build with qemu
binfmt emulation.

* gnu/packages/golang-xyz.scm (go-github-com-hodgesds-perf-utils)[arguments]:
Skip a couple more tests on arm.

Change-Id: Ib5bc41679d825a1a727e240e7fffeb336f82db6d