gnu: knotes: Delete package
* gnu/packages/kde-pim.scm (knotes): Remove variable

The KDE project decided to remove Akonadi Notes.
The Guix package has been failing to build for over two months.

The Guix Deprecation Policy states that software may be removed
immediately if upstream has decided to drop it.
Additionally, if a package fails to build for two months like this one,
it also may be removed immediately.