guix: derivations: Add optional non-distributability marker
This commit introduces the optional `#:distributable?` argument to
`derivation`. Unlike the existing `#:substitutable?` argument, derivations
which have `#:distributable?` set to `#f` will "propagate" their
non-distributability. As a result, derivations will be marked
non-distributable if they have `#:distributable` set to `#f`, or if any input
in their graph of inputs has `#:distributable` set to `#f`.

* guix/derivations.scm (derivation): Introduce `#:distributable?` argument.
* tests/derivations.scm (distributable-derivation?): Test behaviour of changes
to `derivation`.
* doc/guix.texi: Add documentation for new `#:distributable?` argument.
* guix/gexp.scm (gexp->derivation): Introduce `#:distributable?` argument.

Change-Id: I30f6b08ba1fbe5bd7e30084bfa6ff5d4137da004