services: rootless-podman: Make setting / as a shared mount configurable.
Programs using least-authority-wrapper which in turn uses pivot-root to
isolate the containers filesystem. pivot-root crashes in case / is a
shared mount, which rootless containers require. This breaks every
service using least-authority-wrapper on a system where the
rootles-podman-service-type is enabled.

This patch makes the behavior conditional, based on a
rootless-podman-configuration field.

* gnu/services/containers.scm (rootless-podman-configuration): Add
root-share-mount? field.
(rootless-podman-shared-root-fs-entrypoint): If root-shared-mount? is #f
print a message and do not make / a shared mount.
(rootless-podman-shared-root-fs-entrypoint): Use the new field.
* doc/guix.texi: Document it.

Fixes https://codeberg.org/guix/guix/issues/3233

Change-Id: Iaea9975be5b6df8a013a04fde5f836a834920e63