Grub: add lvm module
This allows using LVM2 volumes for things that need to be available
at boot time, like root, /gnu, etc.

Currently, it is necessary to do this manually at each boot because
the code that generates grub.cfg does not recognize or know that
lvm is needed. Adding it is not a big burden, it is a standard
module that is always available.

Change-Id: I44b42f69861e6f1bc4b511286590159cca18ff75