gnu: obs: Disable websocket support.
* gnu/packages/video.scm (obs): Disable websocket support.

Boost deprecated & renamed its `io_service` to something else. websocketpp has
not updated to reflect this fact, making this a known issue now [1]. OBS's
websocket plugin exercises this code-path in websocketpp, causing a
compilation failure. The exact compilation error looks something below:

```
include/websocketpp/transport/asio/connection.hpp:89:24: error: ‘io_service’
in namespace ‘websocketpp::lib::asio’ does not name a type
   89 |     typedef lib::asio::io_service * io_service_ptr;
      |                        ^~~~~~~~~~
```

The fix is available in a PR [2], but it has not been applied yet.

[3] is a Gentoo patch on the mailing list/bug tracker that may fix this, but
it does not apply cleanly. I am leaving this link in case we want to fix this
in websocketpp instead of disabling support in OBS.

[1] https://github.com/zaphoyd/websocketpp/issues/1157
[2] https://github.com/zaphoyd/websocketpp/pull/1164
[3] https://946365.bugs.gentoo.org/attachment.cgi?id=921340

Change-Id: I7103941412dd2d4b2103ab61558bd3a8ae9cfce9