gnu: adwaita-icon-theme: Provide legacy icons.
Adwaita registers itself as an icon theme with an index.theme file, which is
defined as part of the Freedesktop Icon Naming Specification [1].  This
creates an implicit assumption that it conforms to the specification, which
can be utilized by applications.  Recent versions of Adwaita are no longer
compliant, but it still technically advertises itself as such by using this
interface, which means that apps that rely on the FDO specification to acquire
icons may not always be able to find them when Adwaita is the default
theme [2].

The linked issue [2] concluded with the decision to release a legacy version
of adwaita-icon-theme.  This legacy version can be optionally provided by
distributions alongside adwaita-icon-theme to restore compliance with the
specification.

This change seems to only partially fix missing-icons issues with non-GTK apps
on GNOME (on Guix System).  Some apps - e.g., gwenview, partitionmanager -
regain icons, but others - e.g., dolphin, kate - still remain iconless.

[1] https://specifications.freedesktop.org/icon-theme-spec/latest/

[2] https://gitlab.gnome.org/GNOME/adwaita-icon-theme/-/issues/288

* gnu/packages/gnome.scm (adwaita-icon-theme)[propagated-inputs]: Add
adwaita-icon-theme-legacy.

Change-Id: I656826df2adc66e1df5e49c0e480df7fe90df668