ui: Change bug report address to Codeberg.
`guix --help` used to print `Report bugs to: bug-guix@gnu.org.`

This commit changes the message to
`Report bugs to: <https://codeberg.org/guix/guix/issues/>`.

In accordance with the consensus decision
<https://codeberg.org/guix/guix-consensus-documents/src/branch/main/002-codeberg.md>
and just in time before we phase out bug-guix.

* configure.ac (AC_INIT): Set Codeberg as bug-report address, to be used
as %guix-bug-report-address in guix/config.scm.in.
* guix/self.scm (make-config.scm): Likewise, to be applied when
`guix pull` is run a second time.
* guix/ui.scm (show-bug-report-information): Display bug report address
in <> angle brackets.

Change-Id: Ie16dccbf8b9801e278d3aa0dd5eccaab737725bd