daemon: allow ADDR_NO_RANDOMIZE to be omitted if GUIX_ALLOW_ASLR.
Docker's default seccomp filter prevents use of the ADDR_NO_RANDOMIZE flag
with the personality system call.  It causes personality to return EPERM.  In
general, we assume that any result other than the only documented one, EINVAL,
is caused by seccomp.  If we detect that ADDR_NO_RANDOMIZE is blocked, and the
GUIX_ALLOW_ASLR environment variable is set and nonzero, we simply don't use
it.

This allows guix-daemon to continue to work even in these containers, without
any implicit weakening of reproducibility.

* nix/libstore/build.cc (DerivationGoal::startBuilder): detect when
  ADDR_NO_RANDOMIZE is blocked and GUIX_ALLOW_ASLR is set and nonzero and
  don't use it in that case.

Fixes: https://codeberg.org/guix/guix/issues/3917
Change-Id: I51c5899a9559e161f9e107c2e6a36df395ab3134