teams: Add script to generate manifest for operating-system.
Takes an operating-system and generates a manifest out of the
packages that can be found in the records.

This script is not finished completely, it still requires manual
work after generation of the manifest. And it will not generate
a complete manifest! This is because the scripts contain additional
dependencies that aren't captured anywhere lower in the operating-system
record.

* etc/teams/release/os-manifest-generator.scm: Add script.

Change-Id: Ib9a5683a60e7fef1ce049b1c6c5ec4fc77ce6c58