import: npm-binary: Handle `npm:` versions.
I know next to nothing about npm, but I have been seeing an issue where I
cannot import some packages. Noticed it with @isaacs/cliui. This package has
a dependency with a name `wrap-ansi-cjs` and its version is
`npm:wrap-ansi@^7.0.0`. Seems that actual dependency is `wrap-ansi`.
Could someone shed some light onto this, is this a proper solution?

Wasn't able to properly test this as the import of @isaacs/cliui leaves
me with a package that doesn't build, failing to look for string-width that is present
in inputs... and I gave up for now.

* guix/import/npm-binary.scm
(make-versioned-package-inner): Rename constructor of versioned-package.
(make-versioned-package):

Change-Id: I31ee1230a3d33a3fb4a483b72f2b88fe0f20d01a