gnu: Remove store path occurrences.
When running `make dist`, it is checked that there are no store paths in the
resulting dist folder. Specifically through the `assert-no-store-file-names`
target.

I think that this target makes sense and this commit has workarounds for
currently embedded store paths. Two of those are comments. Maybe the target
could be changed to ignore those in the future. But the hash is not important,
so just replace it with <hash>.

As for crate tests, I decided to split the string instead of removing it,
although it seems to me it might safely be removed, but I am not completely
sure about the intent here.

* gnu/packages/astronomy.scm (python-jplephem-2.22): Remove store hash from comment.
* gnu/packages/golang-build.scm (golang-org-x-exp): Remove store hash from comment.
* tests/import/crate.scm (temp-file): Split store path's hash.

Change-Id: I709f29a9dfab626d194c41d12e53e0380a148217