gnu: niri: Restore installation of bin/niri.
Niri package is a workspace, thus it seems it should use #:cargo-install-paths
so that the packages are properly installed. The reason they currently aren't
installed is that manifest-targets looks only on first 'packages', while niri
is third package in the list. The first package is a library, not 'bin', so
there's nothing to install. I am not completely sure this behavior is
intended. And if it is, it seems niri has worked only by chance before.
Ie. if it was first in the list, it would work. In case this is an error
in the cargo-build-system, changing it implies rebuilding all cargo packages,
hence this workaround would still be useful.

Fixes #4321.

* gnu/packages/wm.scm (niri)<arguments>: Add cargo-install-paths.

Change-Id: I7f34f734698f50ef3c913ed858c504b2c99042bf