gnu: mu: Add additional file to patch-bin-references phase.
* gnu/packages/mail.scm (mu): Add mu-utils-file.cc's call to /bin/rm to
patch-bin-references phase.

When mu runs as a server, it creates a randomly-named temporary directory in
`/tmp/`. Mu previously hard-coded a path to `/bin/rm` for removing this
directory when the server closed. We missed this one in the phase that patches
these paths.

This has also been reported upstream, but will only make it into the next
release of mu. In the meantime, we can just patch the /bin/rm for ours and
remove the whole phase when the next release happens.

https://github.com/djcb/mu/issues/2887
https://github.com/djcb/mu/commit/f237a2b9905475fb95da6a04e318d10cab61ddeb
https://github.com/djcb/mu/commit/fc4d5b01a703e8c8cc390cfea135f08d3b45ccab

Change-Id: I419b994ca403a14c20520755116591196b724fae