guix: store: Remove custom port output buffering.
This is mostly motivated by wanting more control over the buffering, which is
the case with this change as it's associated with the port which may be
supplied, rather than added on as a additional layer.

I'm also interested in potentially increasing the buffer size as bigger writes
may improve performance when using inferiors, or when there's high latency on
the connection. This commit doesn't change the buffer size though.

* guix/store.scm (connect-to-daemon): Use block buffering on the returned port
by default.
(port->connection): Don't use the buffering-output-port.
(buffering-output-port): Remove.

Change-Id: Id2c18e2448ae42a8194a0d35c7b1b210a3036df2