gnu: lean4: Fix silent install failure
The lean4 package currently builds and installs without a visible
failure, but fails to effectvely provide its output, such as the "lean"
binary.

This is due to the install phase of the derivation failing silently as
it tries to access a bash shell using an absolute path that expects an
FHS compliant system.

To fix the issue, the relevant path in "src/stdlib.make.in", which is
used during the install phase of the of the project, is now patched out
by the package definition.

* gnu/packages/lean.scm (lean4):
[arguments] Add substitution for FHS path in "src/stdlib.make.in"

Change-Id: Ib3db9ce1fbb46175130f9b46c58c55cd65a4a1ae