import: git: Support 'stable-version-pattern' property.
Some packages (eg: xfce4-panel) use even number for minor version to indicate
a stable release, when refresh such packages to latest versions we'd like them
to be stable ones.  Introduce 'stable-version-pattern' as a package property,
so that unwanted releases can be filtered out.

* doc/guix.texi (Invoking guix refresh): Document 'stable-version-pattern'.
* guix/import/git.scm (import-git-release): Filter out unstable releases when
target-version is unset but 'stable-version-pattern' is available.

Change-Id: I7bbb2cb41b19175a7e3fb79e00cd8dbb20d049ad